+++ breakwater/src/sinks/egui/mod.rs
@@ -2,10 +2,10 @@ use std::{fmt::Display, net::ToSocketAddrs, str::FromStr, sync::Arc};
 
 use async_trait::async_trait;
 use breakwater_parser::FrameBuffer;
+use color_eyre::eyre::{self, Context};
 use dynamic_overlay::UiOverlay;
 use log::error;
 use serde::{Deserialize, Serialize};
-use snafu::{ResultExt, Snafu};
 use tokio::sync::{broadcast, mpsc};
 
 use super::DisplaySink;
@@ -14,18 +14,6 @@ use crate::statistics::StatisticsInformationEvent;
 mod canvas_renderer;
 mod dynamic_overlay;
 mod view;
-
-#[derive(Debug, Snafu)]
-pub enum Error {
-    #[snafu(display("eframe started with an unsupported frontend"))]
-    UnsupportedEguiFrontend,
-
-    #[snafu(display("egui failed"))]
-    EguiFailed,
-
-    #[snafu(display("dynamic overlay error"))]
-    DynamicOverlay { source: dynamic_overlay::Error },
-}
 
 /// Describes the part of the framebuffer that the corresponding viewport will display.
 #[derive(Debug, Serialize, Deserialize, Clone, Copy)]
@@ -59,11 +47,11 @@ impl FromStr for ViewportConfig {
 
         match parts {
             Err(e) => {
-                error!("unable to parse view port config: {e}");
+                error!("failed to parse view port config: {e}");
                 Err(InvalidViewportConfig)
             }
             Ok(parts) if parts.len() != 4 => {
-                error!("unable to parse view port config: invalid format");
+                error!("failed to parse view port config: invalid format");
                 Err(InvalidViewportConfig)
             }
             Ok(parts) => Ok(Self {
@@ -94,7 +82,7 @@ impl<FB: FrameBuffer + Send + Sync + 'static> DisplaySink<FB> for EguiSink<FB> {
         _statistics_tx: mpsc::Sender<crate::statistics::StatisticsEvent>,
         statistics_information_rx: broadcast::Receiver<StatisticsInformationEvent>,
         terminate_signal_rx: broadcast::Receiver<()>,
-    ) -> Result<Option<Self>, super::Error>
+    ) -> eyre::Result<Option<Self>>
     where
         Self: Sized,
     {
@@ -115,7 +103,7 @@ impl<FB: FrameBuffer + Send + Sync + 'static> DisplaySink<FB> for EguiSink<FB> {
 
         let ui_overlay = Arc::new({
             if let Some(ui) = cli_args.ui.as_ref() {
-                dynamic_overlay::load_and_check(ui).context(DynamicOverlaySnafu)?
+                dynamic_overlay::load_and_check(ui).context("failed to load dynamic overlay")?
             } else {
                 UiOverlay::BuiltIn
             }
@@ -149,21 +137,20 @@ impl<FB: FrameBuffer + Send + Sync + 'static> DisplaySink<FB> for EguiSink<FB> {
     }
 
     /// This should only run on the main thread
-    async fn run(&mut self) -> Result<(), super::Error> {
+    async fn run(&mut self) -> eyre::Result<()> {
         // block_in_place below should only be used in a MultiThread runtime
         assert_eq!(
             tokio::runtime::Handle::current().runtime_flavor(),
             tokio::runtime::RuntimeFlavor::MultiThread
         );
 
-        tokio::task::block_in_place(move || self.run_eframe_display())
-            .map_err(|e| {
-                error!("egui failed: {e:?}");
-                snafu::NoneError
-            })
-            .context(EguiFailedSnafu)?;
+        tokio::task::block_in_place(move || {
+            if let Err(e) = self.run_eframe_display() {
+                eyre::bail!("egui failed: {e:?}");
+            }
 
-        Ok(())
+            Ok(())
+        })
     }
 }
 
@@ -196,7 +183,7 @@ impl<FB: FrameBuffer + Send + Sync + 'static> EguiSink<FB> {
                     advertised_endpoints,
                     ui_overlay,
                 )
-                .expect("unable to create egui frontend");
+                .expect("failed to create egui frontend");
 
                 Ok(Box::new(frontend))
             }),